Feedback: Merge cmds.go and main.go into cmd.go
